登录页改为原生，添加推送

+++ app/src/main/java/com/itsv/FSZHZX/model/GroupListModel.java
@@ -124,7 +124,7 @@ public class GroupListModel implements Serializable {
             private String email;
             private String phone;
             private String roleId;
-            private int deptId;
+            private long deptId;
             private String status;
             private String avatar;
             private String position;
@@ -214,11 +214,11 @@ public class GroupListModel implements Serializable {
                 this.roleId = roleId;
             }
 
-            public int getDeptId() {
+            public long getDeptId() {
                 return deptId;
             }
 
-            public void setDeptId(int deptId) {
+            public void setDeptId(long deptId) {
                 this.deptId = deptId;
             }
